chore: make resolveClasses global
